+++ src/vscode/extension.ts
@@ -1,136 +1,29 @@
-import { Graphviz } from "@hpcc-js/wasm-graphviz";
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
-import Parser, { type SyntaxNode } from "web-tree-sitter";
+import type { Language } from "../control-flow/cfg";
 import {
-  type Language,
-  functionNodeTypes,
-  newCFGBuilder,
-} from "../control-flow/cfg";
-import {
-  type CFG,
-  mergeNodeAttrs,
-  remapNodeTargets,
-} from "../control-flow/cfg-defs";
-import {
-  type ColorScheme,
+  type ColorList,
   deserializeColorList,
   getDarkColorList,
   getLightColorList,
-  listToScheme,
 } from "../control-flow/colors";
-import { simplifyCFG, trimFor } from "../control-flow/graph-ops";
-import { graphToDot } from "../control-flow/render";
+import type { UpdateCode, UpdateSettings } from "./messages.ts";
 import { OverviewViewProvider } from "./overview-view";
 
-let graphviz: Graphviz;
-interface SupportedLanguage {
-  /**
-   * The VSCode languageId to match on
-   */
-  languageId: string;
-  /**
-   * The CFG language
-   */
-  language: Language;
-  /**
-   * Name of the parser to use
-   */
-  parserName: string;
-}
 // ADD-LANGUAGES-HERE
-const supportedLanguages: SupportedLanguage[] = [
-  {
-    languageId: "c",
-    language: "C" as Language,
-    parserName: "tree-sitter-c.wasm",
-  },
-  {
-    languageId: "cpp",
-    language: "C++" as Language,
-    parserName: "tree-sitter-cpp.wasm",
-  },
-  {
-    languageId: "go",
-    language: "Go" as Language,
-    parserName: "tree-sitter-go.wasm",
-  },
-  {
-    languageId: "python",
-    language: "Python" as Language,
-    parserName: "tree-sitter-python.wasm",
-  },
-  {
-    languageId: "typescript",
-    language: "TypeScript" as Language,
-    parserName: "tree-sitter-typescript.wasm",
-  },
-  {
-    languageId: "javascript",
-    language: "TypeScript" as Language,
-    parserName: "tree-sitter-typescript.wasm",
-  },
-  {
-    languageId: "typescriptreact",
-    language: "TSX" as Language,
-    parserName: "tree-sitter-tsx.wasm",
-  },
-  {
-    languageId: "javascriptreact",
-    language: "TSX" as Language,
-    parserName: "tree-sitter-tsx.wasm",
-  },
-];
-
-const supportedLanguageIds = new Set(
-  supportedLanguages.map((lang) => lang.languageId),
-);
-const idToLanguage = (languageId: string): Language | null => {
-  for (const lang of supportedLanguages) {
-    if (lang.languageId === languageId) {
-      return lang.language;
-    }
-  }
-  return null;
+const languageMapping: { [key: string]: Language } = {
+  c: "C",
+  cpp: "C++",
+  go: "Go",
+  python: "Python",
+  typescript: "TypeScript",
+  javascript: "TypeScript",
+  typescriptreact: "TypeScript",
+  javascriptreact: "TypeScript",
 };
 
-async function initializeParsers(
-  context: vscode.ExtensionContext,
-): Promise<Record<Language, Parser>> {
-  // const parsers: { [key in Language]: Parser } = {};
-  const parsers: Partial<Record<Language, Parser>> = {};
-
-  for (const lang of supportedLanguages) {
-    const languagePath = vscode.Uri.joinPath(
-      context.extensionUri,
-      "parsers",
-      lang.parserName,
-    ).fsPath;
-    parsers[lang.language] = await initializeParser(context, languagePath);
-  }
-
-  return parsers as Record<Language, Parser>;
-}
-
-async function initializeParser(
-  context: vscode.ExtensionContext,
-  languagePath: string,
-) {
-  await Parser.init({
-    locateFile(_scriptName: string, _scriptDirectory: string) {
-      return vscode.Uri.joinPath(
-        context.extensionUri,
-        "parsers",
-        "tree-sitter.wasm",
-      ).fsPath;
-    },
-  });
-  const parser = new Parser();
-  const Go = await Parser.Language.load(languagePath);
-  parser.setLanguage(Go);
-  return parser;
-}
+const idToLanguage = (languageId: string): Language | undefined => {
+  return languageMapping[languageId];
+};
 
 function getCurrentCode(): {
   code: string;
@@ -144,13 +37,10 @@ function getCurrentCode(): {
 
   const document = editor.document;
   const languageId = document.languageId;
-  if (!supportedLanguageIds.has(languageId)) {
-    console.log(`Unsupported language id: ${languageId}`);
-    return null;
-  }
 
   const language = idToLanguage(languageId);
   if (!language) {
+    console.log(`Unsupported language id: ${languageId}`);
     return null;
   }
 
@@ -167,7 +57,7 @@ type Settings = {
   flatSwitch: boolean;
   simplify: boolean;
   highlightCurrentNode: boolean;
-  colorScheme: ColorScheme;
+  colorList: ColorList;
 };
 type ColorSchemeOptions = "Light" | "Dark" | "Custom" | "System";
 function loadSettings(): Settings {
@@ -200,32 +90,8 @@ function loadSettings(): Settings {
     flatSwitch: config.get("flatSwitch") ?? false,
     simplify: config.get("simplify") ?? false,
     highlightCurrentNode: config.get("highlightCurrentNode") ?? true,
-    colorScheme: listToScheme(colorList),
+    colorList: colorList,
   };
-}
-type CFGKey = { functionText: string; flatSwitch: boolean; simplify: boolean };
-
-function isSameKey(a: CFGKey, b: CFGKey): boolean {
-  return JSON.stringify(a) === JSON.stringify(b);
-}
-
-function getFunctionAtPosition(
-  tree: Parser.Tree,
-  position: vscode.Position,
-  language: Language,
-): Parser.SyntaxNode | null {
-  let syntax: SyntaxNode | null = tree.rootNode.descendantForPosition({
-    row: position.line,
-    column: position.character,
-  });
-
-  while (syntax) {
-    if (functionNodeTypes[language].includes(syntax.type)) {
-      break;
-    }
-    syntax = syntax.parent;
-  }
-  return syntax;
 }
 
 function focusEditor() {
@@ -236,6 +102,7 @@ function focusEditor() {
       preview: false, // Don't open in preview mode
       viewColumn: editor.viewColumn,
     });
+    console.log("Focus!!!");
   }
 }
 
@@ -244,7 +111,7 @@ function moveCursorAndReveal(offset: number) {
   if (!editor) {
     return;
   }
-
+  console.log("Moving!");
   const position = editor.document.positionAt(offset);
   editor.selection = new vscode.Selection(position, position);
 
@@ -258,29 +125,10 @@ function moveCursorAndReveal(offset: number) {
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
 export async function activate(context: vscode.ExtensionContext) {
-  graphviz = await Graphviz.load();
-
-  const helloWorldDark = `digraph G { 
-    bgcolor="#1e1e1e"
-    node [color="#e0e0e0", fontcolor="#e0e0e0"]
-    edge [color="#e0e0e0"]
-    Hello -> World 
-}`;
-  const helloWorldLight = "digraph G { Hello -> World }";
-
-  // We use the color theme for the initial graph, as it's a good way to avoid
-  // shocking the user without being overly complicated with reading custom
-  // color schemes.
-  const helloWorldSvg = graphviz.dot(
-    isThemeDark() ? helloWorldDark : helloWorldLight,
-  );
-  const helloWorldBGColor = isThemeDark() ? "#1e1e1e" : "white";
-
   const provider = new OverviewViewProvider(
     context.extensionUri,
-    helloWorldSvg,
-    helloWorldBGColor,
-    onNodeClick,
+    isThemeDark(),
+    { navigateTo: ({ offset }) => onNodeClick(offset) },
   );
 
   context.subscriptions.push(
@@ -296,14 +144,7 @@ export async function activate(context: vscode.ExtensionContext) {
     'Congratulations, your extension "function-graph-overview" is now active!',
   );
 
-  const parsers = await initializeParsers(context);
-
-  let cfgKey: CFGKey | undefined;
-  let savedCFG: CFG | undefined;
-
-  function onNodeClick(node: string): void {
-    if (!savedCFG) return;
-    const offset = savedCFG.graph.getNodeAttribute(node, "startOffset");
+  function onNodeClick(offset: number): void {
     moveCursorAndReveal(offset);
     focusEditor();
   }
@@ -311,20 +152,15 @@ export async function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.workspace.onDidChangeConfiguration(
       (e: vscode.ConfigurationChangeEvent) => {
-        // TODO: This currently only changes the color-scheme.
-        // TODO: Make this react to all the CFG settings.
         if (e.affectsConfiguration("functionGraphOverview")) {
           const settings = loadSettings();
-          if (!savedCFG) return;
-          const dot = graphToDot(
-            savedCFG,
-            false,
-            undefined,
-            settings.colorScheme,
-          );
-          const svg = graphviz.dot(dot);
-
-          provider.setSVG(svg, settings.colorScheme["graph.background"]);
+          provider.postMessage<UpdateSettings>({
+            tag: "updateSettings",
+            flatSwitch: settings.flatSwitch,
+            simplify: settings.simplify,
+            highlightCurrentNode: settings.highlightCurrentNode,
+            colorList: settings.colorList,
+          });
         }
       },
     ),
@@ -333,11 +169,13 @@ export async function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.window.onDidChangeActiveColorTheme(() => {
       const settings = loadSettings();
-      if (!savedCFG) return;
-      const dot = graphToDot(savedCFG, false, undefined, settings.colorScheme);
-      const svg = graphviz.dot(dot);
-
-      provider.setSVG(svg, settings.colorScheme["graph.background"]);
+      provider.postMessage<UpdateSettings>({
+        tag: "updateSettings",
+        flatSwitch: settings.flatSwitch,
+        simplify: settings.simplify,
+        highlightCurrentNode: settings.highlightCurrentNode,
+        colorList: settings.colorList,
+      });
     }),
   );
 
@@ -356,54 +194,12 @@ export async function activate(context: vscode.ExtensionContext) {
         if (!code || !languageId || !language) {
           return;
         }
-
-        const tree = parsers[language].parse(code);
-
-        const functionSyntax = getFunctionAtPosition(tree, position, language);
-        if (!functionSyntax) return;
-
-        console.log(functionSyntax);
-        const nameSyntax = functionSyntax.childForFieldName("name");
-        if (nameSyntax) {
-          console.log("Currently in", nameSyntax.text);
-        }
-
-        const { flatSwitch, simplify, highlightCurrentNode, colorScheme } =
-          loadSettings();
-        // We'd like to avoid re-running CFG generation for a function if nothing changed.
-        const newKey: CFGKey = {
-          flatSwitch,
-          simplify,
-          functionText: functionSyntax.text,
-        };
-        let cfg: CFG;
-        if (cfgKey && isSameKey(newKey, cfgKey) && savedCFG) {
-          cfg = savedCFG;
-        } else {
-          cfgKey = newKey;
-
-          const builder = newCFGBuilder(language, { flatSwitch });
-          cfg = builder.buildCFG(functionSyntax);
-          cfg = trimFor(cfg);
-          if (simplify) {
-            cfg = simplifyCFG(cfg, mergeNodeAttrs);
-          }
-          cfg = remapNodeTargets(cfg);
-
-          savedCFG = cfg;
-        }
-        // TODO: Highlighting in the DOT is a cute trick, but might become less effective on larger functions.
-        //       So it works for now, but I'll probably need to replace it with CSS so that I only render once per function.
-
-        // Only highlight if there's more than one node to the graph.
-        const shouldHighlight = highlightCurrentNode && cfg.graph.order > 1;
-        const nodeToHighlight = shouldHighlight
-          ? cfg.offsetToNode.get(offset)
-          : undefined;
-        const dot = graphToDot(cfg, false, nodeToHighlight, colorScheme);
-        const svg = graphviz.dot(dot);
-
-        provider.setSVG(svg, colorScheme["graph.background"]);
+        provider.postMessage<UpdateCode>({
+          tag: "updateCode",
+          code,
+          offset,
+          language,
+        });
       },
     ),
   );