+++ CHANGELOG.md
@@ -15,6 +15,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Fixed
 
 - `finally` blocks are now supported in TypeScript. 
+- Methods are now supported in TypeScript.
+
+### Changed
+
+- VSCode and JetBrains now use the same WebView content.
+- Placeholder ("Hello, World!") graph colors are now determined by the color scheme.
+- VSCode settings now applied instantly to the graph.
 
 ## [0.0.13] - 2025-01-06